more type hint changes
