+++ guillotina/catalog/parser.py
@@ -2,7 +2,7 @@ import typing
 from guillotina.utils import get_content_path
 from guillotina.utils import get_content_depth
 from guillotina.catalog.utils import iter_indexes
-from mypy_extensions import TypedDict
+from guillotina.catalog.types import BasicParsedQueryInfo
 
 
 def to_list(value: typing.Union[str, list]) -> typing.List[str]:
@@ -11,17 +11,6 @@ def to_list(value: typing.Union[str, list]) -> typing.List[str]:
     if not isinstance(value, list):
         value = [value]
     return value
-
-
-class BasicParsedQueryInfo(TypedDict):
-    sort_on: typing.Optional[str]
-    sort_dir: typing.Optional[str]
-    from_: int
-    size: int
-    full_objects: bool
-    metadata: typing.Optional[typing.List[str]]
-    excluded_metadata: typing.Optional[typing.List[str]]
-    params: typing.Dict[str, typing.Any]
 
 
 class BaseParser:
