+++ guillotina/catalog/utils.py
@@ -1,8 +1,11 @@
+import logging
 import typing
 
 from guillotina import app_settings
+from guillotina.catalog.types import BasicParsedQueryInfo
 from guillotina.component import get_utilities_for
-from guillotina.component import get_utility, query_multi_adapter
+from guillotina.component import get_utility
+from guillotina.component import query_multi_adapter
 from guillotina.component import query_utility
 from guillotina.content import IResourceFactory
 from guillotina.content import get_all_possible_schemas_for_type
@@ -10,9 +13,9 @@ from guillotina.directives import index_field
 from guillotina.directives import merged_tagged_value_dict
 from guillotina.directives import merged_tagged_value_list
 from guillotina.directives import metadata
-from guillotina.interfaces import ICatalogUtility, ISearchParser
+from guillotina.interfaces import ICatalogUtility
+from guillotina.interfaces import ISearchParser
 from guillotina.utils import execute
-import logging
 
 
 logger = logging.getLogger('guillotina')
@@ -89,7 +92,7 @@ def get_index_definition(name):
         return _cached_indexes[name]
 
 
-def parse_query(context, query, util=None):
+def parse_query(context, query, util=None) -> typing.Optional[BasicParsedQueryInfo]:
     if util is None:
         util = get_utility(ICatalogUtility)
     parser = query_multi_adapter(
